dev-requirements.txt
pytest
pytest-ckan
pytest-cov